Bonds/Settlements refactor, fixes, consistency [GEN-7694] (#434)
* [pipelines] refatoring and fixes to be more structured

docs: bid-distribution README — claim kinds + refreshed usage for current CLI

refactor: decompose generate_bid_settlements_worker into phase functions + DRY cores

C1: extract pure phase functions (build_staker_stake_groups, compute_stake_totals,
marinade_reward_shares, build_result_claims, staker_rewards_breakdown,
split_distributor_fee) with StakeTotals/MarinadeRewards/StakerRewards/FeeSplit structs.

C2: dedupe the two staker-pool loops into distribute_pool + PoolKind, and the four
fee-deposit blocks into split_by_fraction + push_fee_deposit.

Zero behavior change: arithmetic, to_u64 handling, ensure! order/messages,
PriorityFee-before-Bidding emission and ResultSettlementClaims serialization preserved
verbatim. Golden snapshot passes without re-bless; +5 unit tests.

refactor: move two-phase fee optimizer to fee_optimizer.rs; add phase adj_* tests

C3: relocate MAX_ADJ_ITER, BidSettlementValues, BidSettlementTotals,
calculate_bid_settlement_totals and generate_bid_settlements from bidding.rs into a new
generators/fee_optimizer.rs; generate_bid_settlements_worker becomes pub(crate). Import
sites updated in bin/cli.rs and generators/tests.rs. Pure move, zero behavior change.

Adds the previously missing coverage of the bisection result fields: a phase-1-only
scenario asserts adj_min_fee_bps stays at the configured min_fee_bps, and a phase-2
scenario asserts adj_max_fee_bps pins at max_cap while adj_min_fee_bps rises.

Golden snapshot passes without re-bless; 61 tests green.

perf: build StakerStakeGroups once per run instead of per bisect probe

C4: the staker stake-account grouping depends only on the stake metas, not on the fee
caps the bisection mutates, yet was rebuilt on every worker invocation (up to MAX_ADJ_ITER
times per run). Build it once in generate_bid_settlements and pass
&[(&ValidatorSamMeta, Vec<StakerStakeGroup>)] plus epoch into the worker; the worker drops
its stake_meta_index/sam_validator_metas params. filter_map preserves the
"validator with no grouped metas is skipped" semantics.

StakerStakeGroup and build_staker_stake_groups become pub(crate) for fee_optimizer.
Public API unchanged; zero behavior change. Golden snapshot passes without re-bless; 61
tests green.

* ci: add solana-program dep so clippy --all-features compiles the Anchor program

Anchor's idl-build feature (enabled by the org static-analysis clippy --all-features run) makes #[program]/#[derive(Accounts)] reference bare solana_program; declaring it directly resolves the E0433 that was failing the clippy and solana-lints jobs. #448 landed the cargo-deny half but not this.

* [fix] surface per-validator negative realized commissions

* [review] fixes

* [review] copilot points

* dependencies: update

* [review] fixes

* [review] fixes

* [advisory] fixes versions

* [review] fixes

* lockfile: undo accidental crate downgrades, fixes yanked spin 0.9.8

Commit 5054dc0 ("dependencies: update") was generated against a stale
crates.io index and silently downgraded 8 transitive crates below the
versions already on main. Two (anyhow, crossbeam-epoch) were restored in
0d79b29; this restores the remaining six: spin 0.9.9, crossbeam-utils
0.8.22, hybrid-array 0.4.13, libredox 0.1.18, rand 0.10.2,
redox_syscall 0.9.0.

spin 0.9.8 has since been yanked, which is what failed the cargo-deny
advisories check (static-analysis / cargo-deny). The rand 0.10.1
downgrade was also behind the "RUSTSEC-2026-0097 not encountered"
warning against the org deny config ignore list.

* [review] finalizing review points

* [review] coderabbit round: readyz timeout, deterministic institutional order, anomaly logging, api guards

- document StakerPayout stake-basis invariant, test institutional activating == 0
- readyz: bound the DB connectivity check with a 2s timeout (503 on stall)
- institutional settlements: sort by vote_account — HashMap::into_values order is random
- rewards: surface stakers-paid-while-gross-zero anomalies (negative_realized_commissions -> realized_commission_anomalies)
- api: reject --port equal to the internal 9000 listener
- /metrics: emit the Prometheus exposition content type (text/plain; version=0.0.4)
- Marker: TODO note to enforce decoder invariants (deferred)
- tests: find_claim_by_authority split into fee/staker kind-aware helpers
- discord report: unset LC_ALL so LC_NUMERIC=C wins
- http tests: no_gzip() pins client behavior against workspace feature unification
- serde: stake_accounts map values accept number-or-string like the scalar lamport fields
- losslessJson: reject trailing content after the root value (JSON.parse parity);
  destroy the source stream when the parse fails (tokenizer errors never throw
  through write, so the promise rejection now triggers cleanup)
- lossless spec: malformed token after a large valid prefix
- review] rate-limit test: widen burst overshoot to 90 requests
  60 excess tokens need >2s of refill to absorb, making the 429 assertion
robust against slow CI runners without touching production code.

* [review] claude and cleaning

+++ api/src/dto.rs
@@ -86,6 +86,11 @@ pub struct ProtectedEventRecord {
     pub amount: u64,
     #[serde(with = "pubkey_string_conversion")]
     pub vote_account: Pubkey,
+    /// DEPRECATED: the `{ "funder": ... }` wrapper is retained only for backward
+    /// compatibility. The generated settlement JSON now exposes `funder` directly;
+    /// this nested `meta` field will be flattened to a top-level `funder` in a future
+    /// API version. Read `meta.funder` for now.
+    #[schema(deprecated)]
     pub meta: SettlementMeta,
     pub reason: SettlementReason,
 }
@@ -98,12 +103,15 @@ pub struct ValidatorBondRecordSchema {
     vote_account: String,
     authority: String,
     cpmpe: Decimal,
-    max_stake_wanted: f64,
+    // Decimal, not f64: the real DTO (validator_bonds_common::dto::ValidatorBondRecord)
+    // holds rust_decimal::Decimal, which serializes as a JSON *string* — an f64 here
+    // would advertise a lossy number type the API never actually emits.
+    max_stake_wanted: Decimal,
     epoch: u64,
-    funded_amount: f64,
-    effective_amount: f64,
-    remaining_witdraw_request_amount: f64,
-    remainining_settlement_claim_amount: f64,
+    funded_amount: Decimal,
+    effective_amount: Decimal,
+    remaining_witdraw_request_amount: Decimal,
+    remainining_settlement_claim_amount: Decimal,
     #[schema(format = "datetime")]
     updated_at: DateTime<Utc>,
     bond_type: String, // Using String to represent BondType
